requirements.txt
airtable-python-wrapper==0.13.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pip==20.0.2
requests==2.24.0
setuptools==46.1.3
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
